+++ app/Core/Course/Functions/helpers.php
@@ -127,3 +127,26 @@ function pre($value)
     var_dump($value);
     echo '</pre>';
 }
+
+/**
+ * @param $value
+ * @return string
+ */
+function pass_hash(string $value, int $cost = 10): string
+{
+    return password_hash($value, PASSWORD_DEFAULT, ["cost" => $cost]);
+}
+
+/**
+ * @param $value
+ * @param string $hash
+ * @return bool
+ */
+function pass_verify($value, string $hash): bool
+{
+    if (password_verify($value, $hash)) {
+        return true;
+    } else {
+        return false;
+    }
+}