+++ app/Core/Exceptions/templates/error-file.php
@@ -1,0 +1,17 @@
+<body style='background: #F8F8FF;'>
+    <div style='background: #FFF; padding: 15px; font-family: sans-serif;'>
+        <p>
+            <h1 style='color: #EE0000;'>FileSystem alert: error in code execution</h1>
+        </p>
+        <hr>
+        <p><strong>Type error:</strong> Not Found<br>
+            <hr>
+        </p>
+        <p><strong>Code:</strong> 404<br>
+            <hr>
+        </p>
+        <p><strong>Type error:</strong> Folder "<?= $view; ?>" not found<br>
+            <hr>
+        </p>
+    </div>
+</body>