+++ app/Core/Resource/Message.php
@@ -1,30 +1,41 @@
 <?php
 
 namespace Solital\Core\Resource;
-use Solital\Core\Session\Session;
+use Solital\Core\Resource\Session;
 
 class Message 
 {
-    
-    private static $msg;
-    private static $index;
-    
-    public static function newMessage(string $index , string $msg)
+    /**
+     * @param string $index
+     * @param string $msg
+     */
+    public static function new(string $index , string $msg): Message
     {
-        return Session::new($index, $msg);
+        Session::new($index, $msg);
+        return new static;
     }
     
-    public static function getMessage(string $index)
+    /**
+     * @param string $index
+     */
+    public static function get(string $index): string
     {
         if (isset($_SESSION[$index])) {
-            return $_SESSION[$index];
+            return (string)$_SESSION[$index];
         }
     }
     
-    public static function clearMessage(string $index) {
+    /**
+     * @param string $index
+     * @return bool
+     */
+    public static function clear(string $index): bool
+    {
         if (isset($_SESSION[$index])) {
             unset($_SESSION[$index]);
+            return true;
+        } else {
+            return false;
         }
-    }
-    
+    }    
 }